latest2/requirements.txt
tk
Pillow
ultralytics